+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 - Updated `Pcs` trait implementation for Plonky3 v0.4.2 compatibility (#3).
 - Updated Plonky3 dependencies to v0.4.2 (#3).
 - Handle aux boundary values constraints in prover and verifier (#7).
+- Fix panics in verifier (#19).
 
 ## 0.4.0 (2025-12-23)